+++ backend/domain-cvi/src/test/java/com/cvi/image/repository/ImageRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.cvi.image;
+package com.cvi.image.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
